@@ -14,16 +14,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.0",
-    "@angular/common": "^7.1.0",
-    "@angular/compiler": "^7.1.0",
-    "@angular/core": "^7.1.0",
-    "@angular/forms": "^7.1.0",
-    "@angular/http": "^7.1.0",
-    "@angular/platform-browser": "^7.1.0",
-    "@angular/platform-browser-dynamic": "^7.1.0",
-    "@angular/platform-server": "^7.1.0",
-    "@angular/router": "^7.1.0",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@aspnet/signalr": "^1.1.4",
     "@hmcts/frontend": "^0.0.39-alpha",
     "@nguniversal/module-map-ngfactory-loader": "7.1.0",
@@ -50,10 +50,10 @@
     "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.1.0",
+    "@angular-devkit/build-angular": "^0.802.1",
+    "@angular/cli": "^8.3.20",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/applicationinsights-js": "^1.0.9",
     "@types/jasmine": "^2.8.16",
     "@types/jasminewd2": "^2.0.8",
@@ -70,7 +70,7 @@
     "karma-junit-reporter": "^1.2.0",
     "node-sass": "^4.13.0",
     "puppeteer": "^1.20.0",
-    "typescript": "~3.1.1"
+    "typescript": "~3.5.3"
   },
   "optionalDependencies": {
     "protractor": "~5.4.0",
